Creaating package javelin.controller.content
javelin.controller was getting too bloated so refactoring for better navigability.
"Content" is considered anything front-facing that a player might interact with (a new action, new map, etc).
A lot of these blur the line between controller and model, especially when they are serialized due to Java's ease of oding that but conceptually they are not necessarily model (like actions) or need to be serialized (like maps) or could be regenerated and not necessarily persisted (terrain). Some might be more model than controller but that's not going to be looked into right now.